+++ scriptabit/habitica_task.py
@@ -12,7 +12,7 @@ from datetime import datetime
 from tzlocal import get_localzone
 
 from .dates import parse_date_utc
-from .task import CharacterAttribute, Difficulty, Task
+from .task import CharacterAttribute, ChecklistItem, Difficulty, Task
 
 
 class HabiticaTask(Task):
@@ -43,6 +43,16 @@ class HabiticaTask(Task):
 
         if 'attribute' not in task_dict:
             task_dict['attribute'] = CharacterAttribute.default.value
+
+        # The Habitica API chokes if you attempt to update a task with a
+        # checklist in the request data. To work around this we move the
+        # checklist (if any) out of task_dict so it can be handled separately.
+        self.__delete_checklist_items = {}
+        if 'checklist' in task_dict.keys():
+            self.__checklist = task_dict['checklist']
+            del task_dict['checklist']
+        else:
+            self.__checklist = []
 
     @property
     def task_dict(self):
@@ -142,12 +152,44 @@ class HabiticaTask(Task):
             list: The checklist, or an empty list if there are no
                 checklist items.
         """
-        # TODO: implement
-        return []
+        checklist = []
+        for i in self.__checklist:
+            checklist.append(ChecklistItem(
+                name=i['text'],
+                checked=i['completed']))
+
+        return checklist
 
     @checklist.setter
     def checklist(self, checklist):
         """ Sets, or clears the checklist. """
-        # TODO: implement
-        pass
+        new_names = []
+        existing_items = { i['text']: i for i in self.__checklist }
 
+        # check the input list for new and updated items
+        for i in checklist:
+            new_names.append(i.name)
+            if i.name in existing_items:
+                # update existing entry
+                existing_items[i.name]['text'] = i.name
+                existing_items[i.name]['completed'] = i.checked
+            else:
+                # create new item
+                self.__checklist.append({
+                    'text': i.name,
+                    'completed': i.checked })
+
+        # check for deleted items
+        for k,v in existing_items.items():
+            if k not in new_names:
+                self.__delete_checklist_items[k] = v
+
+    @property
+    def raw_checklist(self):
+        """ Gets the raw checklist (JSON dictionary) """
+        return self.__checklist
+
+    @property
+    def checklist_items_to_delete(self):
+        """ Gets the dictionary of checklist items to delete. """
+        return self.__delete_checklist_items