+++ CHANGELOG.rst
@@ -157,3 +157,7 @@ Changelog
 
 * Updated documentation with plugin usage information.
 * Added CSV batch task creation plugin.
+* Slightly improved error handling during task sync. Now an error in a task
+  doesn't bring the whole sync down. Instead it logs the error and skips to the
+  next task.
+* Made sync of task description/extra text optional, with default to False.